Remove RelationRepository
